Fix cannot remove hidden files on linux

+++ tools/SyncFromMainBranch.ps1
@@ -15,25 +15,28 @@ foreach ($SyncPath in $Config.SyncPath)
     Copy-Item -Path $SyncPath -Destination "$TmpFolder/$SyncPath" -Recurse -Force
 }
 
-git config --global user.email "azurepowershell@ms.com"
-git config --global user.name "azurepowershell"
+# git config --global user.email "azurepowershell@ms.com"
+# git config --global user.name "azurepowershell"
 git checkout -b "syncToolsFolder-$BranchName" "origin/$BranchName"
 
 # There are some files or folders who need to be keeped in target branch.
 foreach ($UnSyncPath in $Config.UnSyncPath)
 {
-    Write-Host "Back up $UnSyncPath from $BranchName branch."
-    $ParentFolder = Split-Path -path $UnSyncPath -Parent
-    Copy-Item -Path $UnSyncPath -Destination "$TmpFolder/$ParentFolder" -Recurse -Force
+    if (Test-Path -Path $UnSyncPath)
+    {
+        Write-Host "Back up $UnSyncPath from $BranchName branch."
+        $ParentFolder = Split-Path -path $UnSyncPath -Parent
+        Copy-Item -Path $UnSyncPath -Destination "$TmpFolder/$ParentFolder" -Recurse -Force
+    }
 }
 
 foreach ($SyncPath in $Config.SyncPath)
 {
-    Remove-Item -Path $SyncPath -Recurse
+    Remove-Item -Path $SyncPath -Recurse -Force
     Copy-Item -Path "$TmpFolder/$SyncPath" -Destination $SyncPath -Recurse -Force
-    git add $SyncPath
+    # git add $SyncPath
 }
 
-git commit -m "Sync tools folder from main branch to $BranchName branch"
-git remote set-url origin "https://$GithubToken@github.com/Azure/azure-powershell.git"
-git push origin "syncToolsFolder-$BranchName" --force
+# git commit -m "Sync tools folder from main branch to $BranchName branch"
+# git remote set-url origin "https://$GithubToken@github.com/Azure/azure-powershell.git"
+# git push origin "syncToolsFolder-$BranchName" --force
